fix(esm): break require cycles + netinfo default-import regression
Extract dependency-free leaf modules so cyclic edges no longer form under
Vite/Metro unbundled ESM:

- util/debug-registry: debug-clear registry, split from util/debug (which
  imports the logger). Breaks logger -> ring-logger -> debug -> logger.
- chat/inbox/metadata-store: inbox-metadata store + metasReceived /
  participantInfoReceived, with no router dependency. constants/router imports
  the leaf; metadata.tsx re-exports for other consumers. Breaks the 5
  constants/router <-> chat/inbox/metadata cycles.
- router-v2/deep-link-emitter: emitDeepLink + normalizeUrl + emitter state,
  zero deps. stores/push imports the leaf; linking re-exports. Breaks
  push -> linking cycles.
- stores/push now subscribes to config.loginError to clear its pending push
  notification instead of config calling into push. Breaks config <-> push.

Also fix a require->import ESM regression: @react-native-community/netinfo's
NetInfoStateType is a named export, absent from the default export. Merge it
into the registered native module so NetInfo.NetInfoStateType.none resolves
again (was throwing "Cannot read property 'none' of undefined" at startup).
