+++ shared/constants/init/platform.native.tsx
@@ -1,6 +1,6 @@
 import * as ExpoLocation from 'expo-location'
 import * as ExpoTaskManager from 'expo-task-manager'
-import NetInfo from '@react-native-community/netinfo'
+import NetInfo, {NetInfoStateType} from '@react-native-community/netinfo'
 import {Linking} from 'react-native'
 import {setupAudioMode} from '@/util/audio.native'
 import {requestLocationPermission} from '@/util/platform-specific'
@@ -18,7 +18,10 @@ export const getNative = (): NativeModules =>
     ExpoLocation,
     ExpoTaskManager,
     Linking,
-    NetInfo,
+    // NetInfoStateType is a named export, not part of the default export; merge
+    // it in so consumers can read NetInfo.NetInfoStateType (default-import under
+    // ESM drops named exports that require() used to expose).
+    NetInfo: {...NetInfo, NetInfoStateType},
     androidAppColorSchemeChanged,
     fsCacheDir,
     fsDownloadDir,
