+++ shared/constants/router.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import * as T from './types'
-import {metasReceived, participantInfoReceived, useInboxMetadataState} from '@/chat/inbox/metadata'
+import {metasReceived, participantInfoReceived, useInboxMetadataState} from '@/chat/inbox/metadata-store'
 import {useInboxLayoutState} from '@/chat/inbox/layout-state'
 import {useCurrentUserState} from '@/stores/current-user'
 import type {ThreadInputAction} from '@/chat/conversation/thread-search-route'
@@ -19,7 +19,7 @@ import type {NavigateAppendType, RouteKeys, RootParamList as KBRootParamList} fr
 import type {GetOptionsRet, RouteDef} from './types/router'
 import {isSplit, threadRouteName} from './chat/layout'
 import {ignorePromise, shallowEqual} from './utils'
-import {registerDebugClear} from '@/util/debug'
+import {registerDebugClear} from '@/util/debug-registry'
 import {makeUUID} from '@/util/uuid'
 import * as Meta from './chat/meta'
 import * as Strings from './strings'
