+++ shared/stores/team-building.tsx
@@ -8,7 +8,7 @@ import {mapGetEnsureValue} from '@/util/map'
 import {serviceIdFromString} from '@/util/platforms'
 import {type StoreApi, type UseBoundStore, useStore} from 'zustand'
 import {validateEmailAddress} from '@/util/email-address'
-import {registerDebugClear} from '@/util/debug'
+import {registerDebugClear} from '@/util/debug-registry'
 import {searchWaitingKey} from '@/constants/strings'
 import {navigateUp, getModalStack, navToProfile} from '@/constants/router'
 import {useSettingsContactsState} from '@/stores/settings-contacts'
