+++ shared/util/debug.tsx
@@ -1,29 +1,5 @@
 import logger from '@/logger'
 
-const debugClearCBs = new Array<() => void>()
-const debugUnClearCBs = new Array<() => void>()
-
-export const registerDebugUnClear = (cb: () => void) => {
-  debugUnClearCBs.push(cb)
-}
-export const registerDebugClear = (cb: () => void) => {
-  debugClearCBs.push(cb)
-}
-export const debugClear = __DEV__
-  ? () => {
-      for (const cb of debugClearCBs) {
-        cb()
-      }
-    }
-  : () => {}
-export const debugUnClear = __DEV__
-  ? () => {
-      for (const cb of debugUnClearCBs) {
-        cb()
-      }
-    }
-  : () => {}
-
 // helper to debug method calls into an object
 export function createLoggingProxy<T extends {[key: string]: unknown}>(
   obj: T,
